Move XDR Int dts-dom builder to its own file.
