+++ xdr-transforms/transform.js
@@ -1,90 +1,7 @@
 import * as dom from 'dts-dom';
 import { Property } from 'jscodeshift';
 import fs from 'fs';
-
-function xdrUINT(ns) {
-  const buffer = dom.create.interface('Buffer');
-  const uintInterface = dom.create.interface('UINT');
-
-  uintInterface.members.push(
-    dom.create.property(
-      'MAX_VALUE',
-      dom.type.numberLiteral(4294967295),
-      dom.DeclarationFlags.ReadOnly
-    )
-  );
-  uintInterface.members.push(
-    dom.create.property(
-      'MIN_VALUE',
-      dom.type.numberLiteral(0),
-      dom.DeclarationFlags.ReadOnly
-    )
-  );
-  uintInterface.members.push(
-    dom.create.method(
-      'read',
-      [dom.create.parameter('io', buffer)],
-      dom.type.number
-    )
-  );
-  uintInterface.members.push(
-    dom.create.method(
-      'write',
-      [
-        dom.create.parameter('value', dom.type.number),
-        dom.create.parameter('io', buffer)
-      ],
-      dom.type.void
-    )
-  );
-
-  uintInterface.members.push(
-    dom.create.method(
-      'isValid',
-      [dom.create.parameter('value', dom.type.number)],
-      dom.type.boolean
-    )
-  );
-  uintInterface.members.push(
-    dom.create.method(
-      'toXDR',
-      [dom.create.parameter('value', dom.type.number)],
-      buffer
-    )
-  );
-  uintInterface.members.push(
-    dom.create.method(
-      'fromXDR',
-      [
-        dom.create.parameter('input', buffer),
-        dom.create.parameter(
-          'format',
-          dom.type.stringLiteral('raw'),
-          dom.ParameterFlags.Optional
-        )
-      ],
-      dom.type.number
-    )
-  );
-  uintInterface.members.push(
-    dom.create.method(
-      'fromXDR',
-      [
-        dom.create.parameter('input', dom.type.string),
-        dom.create.parameter(
-          'format',
-          dom.create.union([
-            dom.type.stringLiteral('hex'),
-            dom.type.stringLiteral('base64')
-          ])
-        )
-      ],
-      dom.type.number
-    )
-  );
-
-  return uintInterface;
-}
+import xdrInt from './unsigned-integer';
 
 export default function transformer(file, api) {
   const j = api.jscodeshift;
@@ -93,8 +10,11 @@ export default function transformer(file, api) {
 
   const ns = dom.create.namespace('xdr');
 
-  const uintInterface = xdrUINT(ns);
-  ns.members.push(uintInterface);
+  const signedInt = xdrInt(ns, 'INT', Math.pow(2, 31) - 1, -Math.pow(2, 31));
+  const unsignedInt = xdrInt(ns, 'UINT', Math.pow(2, 32) - 1, 0);
+
+  ns.members.push(signedInt);
+  ns.members.push(unsignedInt);
 
   xdrDefs.find(types.namedTypes.CallExpression).forEach((p) => {
     const node = p.value;
